+++ changelog.md
@@ -1,2 +1,1 @@
-- Add Simplified Chinese translation (Andypsl8)
-- Fix config on recent NeoForge versions
+- Add unrecognized message feedback